requirements.txt
google-genai>=1.41.0
playwright>=1.55.0
python-dotenv>=1.1.0
